Update and bug fixes, main fix Search for animals
A working WPF program in C#, with the functions of registering an animal, viewing a list, searching for an animal using many criteria, saving animals to a local database. Future possible functions: editing, deleting information about animals, the ability to leave more information and view this information in a separate window.