Several edits and addition of new features.
A window for detailed viewing of information about the animal has been added, in which the main function is 'Notes', in which you can write and edit information as you wish. The ability to edit such information as 'Years of life', 'Health status', 'Short description' has been added. The ability to delete an animal from the list and local database has been added. The function of selecting an animal in the animal search window has been fixed. Some minor bugs have also been fixed.